code_examples/python/rag/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.8
decorator==5.1.1
distro==1.9.0
executing==2.1.0
faiss-cpu==1.9.0
frozenlist==1.5.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter-client==8.6.3
jupyter-core==5.7.2
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.17
langchain-ollama==0.2.0
langchain-openai==0.2.3
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.142
marshmallow==3.23.1
matplotlib-inline==0.1.7
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
ollama==0.3.3
openai==1.52.2
orjson==3.10.11
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pinecone==5.3.1
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pip==24.3.1
platformdirs==4.3.6
prompt-toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.6.1
pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.35
stack-data==0.6.3
tenacity==9.0.0
tiktoken==0.8.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
yarl==1.17.1
